+++ src/main/java/com/nexus/event/EventRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface EventRepository extends JpaRepository<Event, Integer> {
+public interface EventRepository extends JpaRepository<Event, Long> {
     @Query("""
         SELECT e FROM Event e
         join fetch e.admins a
@@ -23,6 +23,6 @@ public interface EventRepository extends JpaRepository<Event, Integer> {
     SET e.urgent = true
     WHERE e.id IN :ids
     """)
-    void updateUrgentToTrue(@Param("ids") Iterable<Integer> ids);
+    void updateUrgentToTrue(@Param("ids") Iterable<Long> ids);
 
 }
