+++ src/main/java/com/nexus/expense/ExpenseResponse.java
@@ -3,7 +3,7 @@ package com.nexus.expense;
 import com.nexus.expensecategory.BasicExpenseCategoryResponse;
 
 public record ExpenseResponse(
-        Integer id,
+        long id,
         double amount,
         String paymentDate,
         BasicExpenseCategoryResponse category
