+++ src/main/java/com/nexus/expensecategory/ExpenseCategoryRepository.java
@@ -2,5 +2,5 @@ package com.nexus.expensecategory;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExpenseCategoryRepository extends JpaRepository<ExpenseCategory, Integer> {
+public interface ExpenseCategoryRepository extends JpaRepository<ExpenseCategory, Long> {
 }
